Add unset fonction
Changes:
- unset function;
- env utils with env loopuk by key;
- todo: unset should not only remove but also shrink env array;

Task: https://github.com/a3nv/42prague_minishell/issues/6
