+++ src/executor/builtin_export_utils.c
@@ -6,7 +6,7 @@
 /*   By: iasonov <iasonov@student.42prague.com>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/25 17:47:56 by iasonov           #+#    #+#             */
-/*   Updated: 2024/12/26 23:56:10 by iasonov          ###   ########.fr       */
+/*   Updated: 2024/12/27 20:50:46 by iasonov          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,21 +35,26 @@ void	clear_copy(char **copy)
 	free(copy);
 }
 
-/**
- * todo: both key an value might be NULL handle this
- */
-t_pair	*parse_arg(char *arg)
+char	**copy_envp(char **original)
 {
-	char	*separator_pt;
-	long	key_index;
-	char	*key;
-	char	*value;
+	int		count;
+	int		i;
+	char	**nenvp;
 
-	separator_pt = ft_strchr(arg, '=');
-	if (!separator_pt)
+	count = count_env_vars(original);
+	nenvp = malloc((count + 2) * sizeof(char *));
+	if (!nenvp)
 		return (NULL);
-	key_index = separator_pt - arg;
-	key = ft_strndup(arg, key_index);
-	value = ft_strdup(separator_pt + 1);
-	return (create_pair(key, value));
+	i = 0;
+	while (original[i])
+	{
+		nenvp[i] = ft_strdup(original[i]);
+		if (!nenvp[i])
+		{
+			clear_copy(nenvp);
+			return (NULL);
+		}
+		i++;
+	}
+	return (nenvp);
 }
